+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -1292,6 +1292,12 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
       run("StaticGraphCompanionsSeparateModules.kt");
     }
 
+    @Test
+    @TestMetadata("TwoScopesWithTheSameName.kt")
+    public void testTwoScopesWithTheSameName() {
+      run("TwoScopesWithTheSameName.kt");
+    }
+
     @Nested
     @TestMetadata("compiler-tests/src/test/data/box/dependencygraph/bindingcontainers")
     @TestDataPath("$PROJECT_ROOT")