+++ CHANGELOG.md
@@ -70,6 +70,7 @@ Example rich output (note that in rich terminals this would have color and marku
 
 ### Fixes
 
+- **[FIR/IR]** Fully qualify generated hint function names so scope classes with the same simple name don't collide.
 - **[IR]** Preserve substituted generic type arguments when generating assisted factory delegate parameters and dynamic graph container inputs.
 - **[IR]** Forward extension and context receivers when generated binding-container factories invoke the original binding function.
 - **[IR]** Fix dispatch receivers for generated graph factory functions and companion/object factory accessors.
@@ -81,6 +82,13 @@ Example rich output (note that in rich terminals this would have color and marku
 ### Changes
 
 - Run Metro's functional compiler unit tests on JS.
+
+### Contributors
+
+Special thanks to the following contributors for contributing to this release!
+
+- [@BraisGabin](https://github.com/BraisGabin)
+- [@ychescale9](https://github.com/ychescale9)
 
 ### [Consider sponsoring Metro's development](https://www.zacsweers.dev/sponsoring-metro/)